refactor: share firebase admin
