+++ frontend/src/pages/api/credit-user.js
@@ -1,4 +1,4 @@
-import { db } from '../../lib/firebaseAdmin';
+import { db } from '../../../../shared/firebaseAdmin';
 
 export default async function handler(req, res) {
   if (req.method !== 'POST') {
