+++ frontend/src/pages/api/credits.js
@@ -1,4 +1,4 @@
-import { db } from '../../lib/firebaseAdmin';
+import { db } from '../../../../shared/firebaseAdmin';
 
 export default async function handler(req, res) {
   const { uid } = req.query;
